// projects/ui-builder-component-lib/src/lib/components/map/enums/map-action-types.enum.ts
export enum MapActionTypes {
  MapLoaded = 'map-loaded',
  MarkerClick = 'marker-click',
  MarkerUpdated = 'marker-updated',
  MapsAlertsEvent = 'maps-alerts-event',
  UpdatedLocations = 'updated-locations',
  SetMapsSearch = 'set-maps-search',
  SetDeleteLocationIndex = 'set-delete-location-index',
  SetEditLocationIndex = 'set-edit-location-index',
  MapClick = 'map-click',
  MapCenterChanged = 'map-center-changed',
}